+++ lib/services/customers_service.rb
@@ -31,6 +31,27 @@ module Payrex
         )
       end
 
+      def list_payment_methods(id, payload = {})
+        request(
+          is_list: true,
+          method: :get,
+          object: Payrex::Entities::PaymentMethod,
+          path: "#{PATH}/#{id}/payment_methods",
+          payload: payload
+        )
+      end
+
+      def delete_payment_method(id, payload = {})
+        payment_method_id = payload["payment_method_id"] || payload[:payment_method_id]
+
+        request(
+          method: :delete,
+          object: Payrex::Entities::Deleted,
+          path: "#{PATH}/#{id}/payment_methods/#{payment_method_id}",
+          payload: {}
+        )
+      end
+
       def update(id, payload)
         request(
           method: :put,